Added a menu on the side to add buttons
Christiaan and I made it possible to add buttons to the side menu. This is done by putting it in the MenuButton table. This table then gets added to the panel of the menu and this is then added to our entire frame. The buttons itself don't do anything yet and their proper image is not yet correct, but this is something for later.